requirements.txt
langchain
langchain-google-genai
langgraph
python-dotenv
langchain_core
pydantic
pillow